+++ server/src/main/java/com/services/MovieService.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public interface MovieService {
     MovieDto saveMovie(MovieDto movieDto);
-    MovieDto getMovieById(Long id);
+    MovieDto getMovieById(int id);
     MovieDto getMovieByTitle(String title);
     List<MovieDto> getAllMovies();
-    MovieDto updateMovie(Long id, MovieDto movieDto);
-    boolean deleteMovie(Long id);
+    MovieDto updateMovie(int id, MovieDto movieDto);
+    boolean deleteMovie(int id);
 }
